Fix build bug:
new .swift files need to be registered in the .xcodeproj to be compiled
  - dropping a file in the directory isn't enough.

By moving the ReasonsISFSchedule struct into Autotune.swift (which is already in the build target),
all five "cannot find type 'ReasonsISFSchedule' in scope" errors are resolved with no project file
changes needed.
